Don't import argparse in toplevel __init__.py
