+++ pycbc/tmpltbank/option_utils.py
@@ -18,8 +18,9 @@ import argparse
 import logging
 import textwrap
 from pycbc.tmpltbank.lambda_mapping import *
-from pycbc import pnutils, positive_float, nonnegative_float
+from pycbc import pnutils
 from pycbc.tmpltbank.em_progenitors import load_ns_sequence
+from pycbc.types import positive_float, nonnegative_float
 
 class IndentedHelpFormatterWithNL(argparse.ArgumentDefaultsHelpFormatter):
     """
